Update main.py Made function list loadable from a saved state ...
Overall this important update allows you to save your automations and then load them up again. (Python does not allow saving direct function calls of a class because those references get broken once the program is closed and reopened). Now this update saves a list of function descriptions and arguments (funcOutlineList). This does not have direct references to the functions but only names the functions as strings. When you open the file of Automations this list of descriptions is used to build the real list of function calls (funcList). This funcList is then used to call the functions.

The Pickle functions are also separted into save and load