Update TKUserGuidedAutomation.py Ver 1.2 Fixes Save File
Fixed the problem with saving files. Now all past automations are saved. The list (automationObjList) is now the loaded at the outset of running the program so that it gets all previous saved automations before starting. The pickle loading phase is moved to the TK main function. Also a slight rename in the TK Component.