Update TKUserGuidedAutomation.py v1.2 Important Fix of Accessing runAutomation
Without realizing it although each button for each automated activity was getting the correct name (eg: open browser) it was not actually running the correct automation functions. Rather each button was running the same automation rather than its correct one. This problem was related to using lambda in the command option of the Button class. I took lambda out and the problem is fixed:

    def loadButtons(self):
        for object in self.automationObjList: # take each Automation object and instantiate a Button
            self.buttonList.append(Button(self.frame1, text=object.getName(), width=12, command=object.runAutomation))
        for button in self.buttonList: # for each button pack it
            button.pack()