Update TKUserGuidedAutomation.py Ver 1.3 Right Click Menu Added
Added a right click menu which works on each automation button allowing you to delete that given automation. It was also neccessary to clear the buttonList when deleting a button:
1. delete the button from the automationObjList
2. Clear the list self.buttonsList

Note: the button.bind needs to be handled carefully so that each button has a clear association to its attributes when calling the delete button. This way it deletes the correct button according to the attributes of that button.