Update MainFuncsUserGuidAuto.py ver. 1.31
Fixed run automation process which would cause duplicate and triplicate runs etc within a given session. Now the buildActualFubctions() method from AutomationSet checks if the actualFunctions list already has its functions, if so it does not rebuild them (which actually appends them to the previous build) if it does not have the functions only in this case will it proceed to build them.